VALGOL I compiler based on META II spec is failing
The start of the following block expression parses the `.` following
`X` as the `IDSEQ` keyword that means "parse another IDSEQ1".

    ```
    .BEGIN
    .REAL X .,
    ```

Relevant parts of the VALGOL I grammar include:

    ```
    IDSEQ1 = .ID .LABEL * .OUT('BLK 1') .,

    IDSEQ = IDSEQ1 $('.' IDSEQ1) .,

    DEC = '.REAL' .OUT('B  ' *1) IDSEQ .LABEL *1 .,

    BLOCK = '.BEGIN' (DEC '.,' / .EMPTY)
         ST $('.,' ST) '.END' .,
    ```

Intuitively I'd expect the `.` to be the `.,` terminating the `DEC`
portion of the `BLOCK`, but it seems perfectly valid to parse the '.'
and fail when `,` doesn't parse as an `IDSEQ1`.

I think the META II machine and compiler are behaving as expected,
although I may have missed a backtracking mechanism in META II
repetition sequences ('$'). A fix for such a bug would go in the
hand-compiled META II compiler.