Adapt example VALGOL I program to be consistent with compiler output
These are minor syntactic concerns which have no impact on semantics.

The `BLK` ID of `001` clearly doesn't match the grammar. Favoring the
grammar.

Arguably the `EDT` change depends on how we decide to encode literal
strings. I'm opting to skip the length prefix because it's not used in
META II nor the high-level VALGOL I example.